fix the bug of new product

+++ GALNewGUI/Product/EditProduct.xaml.cs
@@ -1,6 +1,7 @@
 ﻿using GALNewGUI.Controls;
 using GALNewGUI.Entity;
 using GALNewGUI.JsonEncryption;
+using Microsoft.Win32;
 using Newtonsoft.Json;
 using System;
 using System.Collections.Generic;
@@ -31,7 +32,9 @@ namespace GALNewGUI.Product
         private static readonly Regex _regex = new Regex(@"^[0-9]*(\.[0-9]*)?$");
         private string productFilePath;
         private Root root = null;
-        //private ProductParameters productParameter;
+        private bool flag = false;
+        private bool isSaveFile = false;
+        private ProductParameters productParameter=null;
         public EditProduct()
         {
             InitializeComponent();
@@ -61,7 +64,7 @@ namespace GALNewGUI.Product
                 e.CancelCommand();
             }
         }
-        public EditProduct(string filePath)
+        public EditProduct(string filePath,bool f)
         {
             InitializeComponent();
             productFilePath = filePath;
@@ -71,6 +74,17 @@ namespace GALNewGUI.Product
             
             }
             LoadProduct(productFilePath);
+            flag = f;
+        }
+        private bool AreObjectsEqualByValue<T>(T obj1, T obj2)
+        {
+            if (obj1 == null && obj2 == null) return true;
+            if (obj1 == null || obj2 == null) return false;
+
+            var json1 = JsonConvert.SerializeObject(obj1);
+            var json2 = JsonConvert.SerializeObject(obj2);
+
+            return json1 == json2;
         }
         private void CloseButton_Click(object sender, RoutedEventArgs e)
         {
@@ -78,20 +92,45 @@ namespace GALNewGUI.Product
         }
         private void EditProduct_Closing(object sender, CancelEventArgs e)
         {
-            MessageBoxResult result = MessageBox.Show("Do you want to save the product file before exiting this page?","Confirm Exit",MessageBoxButton.YesNo,MessageBoxImage.Question);
 
-            if (result != MessageBoxResult.Yes)
+            if (root.ProductParameters.LeftTable == null)
             {
-                //this.Close();
+                root.ProductParameters.LeftTable = new LeftTable();
+                root.ProductParameters.LeftTable.Items = ProductHierarchyViewLeftTable.GetCurrentTableData();
             }
             else
             {
-                SaveProductFile();
-                MessageBox.Show("Save Product File Successfully");
+
+                root.ProductParameters.LeftTable.Items = ProductHierarchyViewLeftTable.GetCurrentTableData();
+            }
+            if (root.ProductParameters.RightTable == null)
+            {
+                root.ProductParameters.RightTable = new RightTable();
+                root.ProductParameters.RightTable.Items = ProductHierarchyViewRightTable.GetCurrentTableData();
+            }
+            else
+            {
+                root.ProductParameters.RightTable.Items = ProductHierarchyViewRightTable.GetCurrentTableData();
+            }
+            if (!AreObjectsEqualByValue<ProductParameters>(root.ProductParameters,productParameter))
+            {
+                MessageBoxResult result = MessageBox.Show("Do you want to save the product file before exiting this page?", "Confirm Exit", MessageBoxButton.YesNo, MessageBoxImage.Question);
+
+                if (result != MessageBoxResult.Yes)
+                {
+                    //this.Close();
+                }
+                else
+                {
+                    SaveProductFile();
+                  
+                }
             }
         }
+
         private void LoadProduct(string filePath)
         {
+
             JsonMechanism jsonDecryption = new JsonMechanism();
             string json = string.Empty;
             if (File.Exists(filePath))
@@ -124,11 +163,50 @@ namespace GALNewGUI.Product
             {
                 root = new Root();
                 root.ProductParameters = new ProductParameters();
+                root.ProductParameters.InputTray = new InputTray();
+                root.ProductParameters.InputTray.ModuleBarcodePoints = new ModuleBarcodePoints();
+                root.ProductParameters.InputTray.BarcodeOffsetPoints = new BarcodeOffsetPoints();
+                root.ProductParameters.Options = new Options();
+                root.ProductParameters.AdaptorPallet = new AdaptorPallet();
+                root.ProductParameters.GALTimer = new GALTimer();
+                root.ProductParameters.LeftTable = new LeftTable();
+                root.ProductParameters.RightTable = new RightTable();
+                root.ProductParameters.AdaptorPallet.PlacementOffset = new PlacementOffset();
+                root.ProductParameters.AdaptorPallet.BarcodeOffset_Points = new BarcodeOffsetPoints();
                 this.DataContext = root.ProductParameters;
-            }  
+
+            }
+
+            var tempObj = JsonConvert.SerializeObject(root.ProductParameters);
+            productParameter = JsonConvert.DeserializeObject<ProductParameters>(tempObj);
+            //productParameter = root.ProductParameters.ShallowCopy();
+            
+            //productParameter.InputTray = root.ProductParameters.InputTray.ShallowCopy();
+            //productParameter = root.ProductParameters.ShallowCopy();
+            //productParameter.InputTray = root.ProductParameters.InputTray.ShallowCopy();
+            //productParameter.InputTray.BarcodeOffset_Points = root.ProductParameters.InputTray.BarcodeOffset_Points.ShallowCopy();
+            //productParameter.InputTray.ModuleBarcode_Points = root.ProductParameters.InputTray.ModuleBarcode_Points.ShallowCopy();
+            //productParameter.InputTray.CheckSpot1Offset = root.ProductParameters.InputTray.CheckSpot1Offset.ShallowCopy();
+            //productParameter.InputTray.CheckSpot2Offset = root.ProductParameters.InputTray.CheckSpot2Offset.ShallowCopy();
+            //productParameter.InputTray.CheckSpot3Offset = root.ProductParameters.InputTray.CheckSpot3Offset.ShallowCopy();
+            //productParameter.InputTray.CheckSpot4Offset = root.ProductParameters.InputTray.CheckSpot4Offset.ShallowCopy();
+            //productParameter.InputTray.CheckSpotEnable = root.ProductParameters.InputTray.CheckSpotEnable.ShallowCopy();
+            //productParameter.InputTray.CheckSpotPosOnOff = root.ProductParameters.InputTray.CheckSpotPosOnOff.ShallowCopy();
+            //productParameter.InputTray.FlyCheckSpot1Offset = root.ProductParameters.InputTray.FlyCheckSpot1Offset.ShallowCopy();
+            //productParameter.InputTray.FlyCheckSpot2Offset = root.ProductParameters.InputTray.FlyCheckSpot2Offset.ShallowCopy();
+            //productParameter.InputTray.FlyCheckSpot3Offset = root.ProductParameters.InputTray.FlyCheckSpot3Offset.ShallowCopy();
+            //productParameter.InputTray.FlyCheckSpot4Offset = root.ProductParameters.InputTray.FlyCheckSpot4Offset.ShallowCopy();
+            //productParameter.InputTray.PlacementOffset = root.ProductParameters.InputTray.PlacementOffset.ShallowCopy();
+
+            //productParameter.Options = root.ProductParameters.Options.ShallowCopy();
+            //productParameter.AdaptorPallet = root.ProductParameters.AdaptorPallet.ShallowCopy();
+            //productParameter.GALTimer = root.ProductParameters.GALTimer.ShallowCopy();
+            //productParameter.LeftTable = root.ProductParameters.LeftTable.ShallowCopy();
+            //productParameter.RightTable = root.ProductParameters.RightTable.ShallowCopy();
         }
         private void SaveProductFile()
         {
+            string fileName = string.Empty;
             if (root.ProductParameters.LeftTable == null)
             {
                 root.ProductParameters.LeftTable = new LeftTable();
@@ -152,43 +230,54 @@ namespace GALNewGUI.Product
 
             string json = JsonConvert.SerializeObject(root);
             json = jsonEncryption.EncryptionString(json);
-            if (File.Exists(productFilePath))
+            if (flag)
             {
-                File.Delete(productFilePath);
-                using (StreamWriter writer = new StreamWriter(productFilePath))
+                if (File.Exists(productFilePath))
                 {
-                    writer.WriteLine(json);
-                }
+                    File.Delete(productFilePath);
+                    using (StreamWriter writer = new StreamWriter(productFilePath))
+                    {
+                        writer.WriteLine(json);
+                    }
 
+                }
+                else
+                {
+
+                    File.WriteAllText(productFilePath, json);
+                }
+                MessageBox.Show("Save Product File Successfully");
             }
             else
             {
+                SaveFileDialog openFileDialog = new SaveFileDialog();
+                openFileDialog.InitialDirectory = @"C:\router\Product File";
+                openFileDialog.Filter = "JSON files (*.json)|*.json|All files (*.*)|*.*";
+                fileName = System.IO.Path.GetFileName(productFilePath);
+                openFileDialog.FileName = fileName;
 
-                File.WriteAllText(productFilePath, json);
+                bool? result = openFileDialog.ShowDialog();
+                if (result.Value == true)
+                {
+                    File.WriteAllText(openFileDialog.FileName, json);
+                    MessageBox.Show("Save Product File Successfully");
+                }
             }
+            var tempObj = JsonConvert.SerializeObject(root.ProductParameters);
+            productParameter = JsonConvert.DeserializeObject<ProductParameters>(tempObj);
 
         }
         protected void Close_Click(object sender, RoutedEventArgs e)
         {
-            //MessageBoxResult result = MessageBox.Show("Do you want to save the product file before exiting this page?", "Confirm Exit", MessageBoxButton.YesNo, MessageBoxImage.Question);
-
-            //if (result != MessageBoxResult.Yes)
-            //{
-            //    this.Close();
-            //}
-            //else
-            //{
-            //    SaveProductFile();
-            //    MessageBox.Show("Save Product File Successfully");
-            //}
             this.Close();
         }
         protected void SaveFile_Click(object sender, RoutedEventArgs e)
         {
             if (root != null)
             {
+               
                 SaveProductFile();
-                MessageBox.Show("Save Product File Successfully");
+              
             }
         
         }
